requirements.txt
Flask==2.2.5
flask-cors==3.0.10
gunicorn==20.1.0
tensorflow-cpu==2.9.3
numpy==1.23.5
Werkzeug==2.2.3
